community-spark/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# LangGraph and LangChain
langgraph>=0.0.1
langchain>=0.1.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0

# WebAuthn
webauthn>=2.0.0

# UI
streamlit>=1.28.0
requests>=2.31.0

# Visualization (optional)
graphviz>=0.20.0

# OpenAI (optional - for LLM-enhanced agent reasoning)
openai>=1.0.0

# Development (optional)
# pytest>=7.4.0
# black>=23.0.0
# ruff>=0.1.0

